# 8260617: Merge ZipFile encoding check with the initial hash calculation