@cl4es 2021-02-02
Added some comments and refactored slightly after offline feedback from Eirik.

/reviewer add @eirbjo